ocrproject/requirements.txt
Flask==2.3.3
opencv-python==4.9.0.80
numpy==1.24.3
Pillow==10.2.0
pdf2image==1.17.0
pytesseract==0.3.10
easyocr==1.7.1
paddleocr==2.8.1
paddlepaddle-tiny==2.6.1

# Optional: faster PDF rendering (requires poppler)
poppler-utils==0.0.0   # This is just a marker — install poppler separately